# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Secrets API
  version: v1
  description: Lightweight secrets manager for simplicity and security.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/secrets:
    get:
      tags: [secrets]
      summary: List secrets
      security:
        - bearerAuth: []
      parameters:
        - name: after_path
          in: query
          description: Cursor for pagination (path of last item from previous page)
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of items to return (clamped to 1000)
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
      responses:
        "200":
          description: Secrets list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSecretsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/secrets/{path}:
    parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [secrets]
      summary: Create or update secret
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretWriteRequest"
      responses:
        "201":
          description: Secret written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretWriteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    get:
      tags: [secrets]
      summary: Read latest secret
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Secret value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretReadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [secrets]
      summary: Delete latest secret version
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds to wait before retrying the request
          schema:
            type: string
            example: "1"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: rate_limit_exceeded
            message: Too many requests. Please retry after the specified delay.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    SecretWriteRequest:
      type: object
      properties:
        value:
          type: string
          description: Base64-encoded plaintext
      required: [value]
    SecretWriteResponse:
      type: object
      properties:
        id:
          type: string
        path:
          type: string
        version:
          type: integer
        created_at:
          type: string
          format: date-time
    SecretReadResponse:
      type: object
      properties:
        id:
          type: string
        path:
          type: string
        version:
          type: integer
        value:
          type: string
          description: Base64-encoded plaintext
        created_at:
          type: string
          format: date-time
...
    ListSecretsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SecretWriteResponse"
        next_cursor:
          type: string
          description: Cursor for the next page (present only if more results exist)
      required: [data]
...