# internal/secrets/http/dto/list_secrets_response.go
// Package dto provides data transfer objects for HTTP request and response handling.
package dto

import (
	secretsDomain "github.com/allisson/secrets/internal/secrets/domain"
)

// ListSecretsResponse represents a paginated list of secrets in API responses.
type ListSecretsResponse struct {
	Data       []SecretResponse `json:"data"`
	NextCursor *string          `json:"next_cursor,omitempty"`
}

// MapSecretsToListResponse converts a slice of domain secrets to a list response.
func MapSecretsToListResponse(secrets []*secretsDomain.Secret, nextCursor *string) ListSecretsResponse {
	data := make([]SecretResponse, 0, len(secrets))
	for _, secret := range secrets {
		data = append(data, MapSecretToCreateResponse(secret))
	}

	return ListSecretsResponse{
		Data:       data,
		NextCursor: nextCursor,
	}
}