# internal/secrets/http/secret_handler.go
// Package http provides HTTP handlers for secret management operations.
// Secrets are encrypted at rest using envelope encryption and can be versioned.
package http

import (
	"encoding/base64"
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"strconv"
	"strings"

	"github.com/gin-gonic/gin"

	authUseCase "github.com/allisson/secrets/internal/auth/usecase"
	cryptoDomain "github.com/allisson/secrets/internal/crypto/domain"
	"github.com/allisson/secrets/internal/httputil"
	secretsDomain "github.com/allisson/secrets/internal/secrets/domain"
	"github.com/allisson/secrets/internal/secrets/http/dto"
	secretsUseCase "github.com/allisson/secrets/internal/secrets/usecase"
	customValidation "github.com/allisson/secrets/internal/validation"
)

// SecretHandler handles HTTP requests for secret management operations.
// It coordinates authentication, authorization, and audit logging with the SecretUseCase.
type SecretHandler struct {
	secretUseCase   secretsUseCase.SecretUseCase
	auditLogUseCase authUseCase.AuditLogUseCase
	logger          *slog.Logger
}

// NewSecretHandler creates a new secret handler with required dependencies.
func NewSecretHandler(
	secretUseCase secretsUseCase.SecretUseCase,
	auditLogUseCase authUseCase.AuditLogUseCase,
	logger *slog.Logger,
) *SecretHandler {
	return &SecretHandler{
		secretUseCase:   secretUseCase,
		auditLogUseCase: auditLogUseCase,
		logger:          logger,
	}
}

// extractPath extracts and validates the path parameter from the request.
// Returns the cleaned path and true if valid, empty string and false if invalid.
func (h *SecretHandler) extractPath(c *gin.Context) (string, bool) {
	path := strings.TrimPrefix(c.Param("path"), "/")
	if path == "" {
		httputil.HandleBadRequestGin(
			c,
			errors.New("path cannot be empty"),
			h.logger,
		)
		return "", false
	}
	return path, true
}

// CreateOrUpdateHandler creates a new secret or updates an existing one.
// POST /v1/secrets/*path - Requires EncryptCapability.
// Returns 201 Created with secret metadata (excludes plaintext value for security).
func (h *SecretHandler) CreateOrUpdateHandler(c *gin.Context) {
	// Extract and validate path from URL parameter
	path, ok := h.extractPath(c)
	if !ok {
		return
	}

	var req dto.CreateOrUpdateSecretRequest

	// Parse and bind JSON
	if err := c.ShouldBindJSON(&req); err != nil {
		httputil.HandleBadRequestGin(c, err, h.logger)
		return
	}

	// Validate request
	if err := req.Validate(); err != nil {
		httputil.HandleValidationErrorGin(c, customValidation.WrapValidationError(err), h.logger)
		return
	}

	// Decode base64 value
	value, err := base64.StdEncoding.DecodeString(req.Value)
	if err != nil {
		httputil.HandleBadRequestGin(
			c,
			fmt.Errorf("invalid base64 value: %w", err),
			h.logger,
		)
		return
	}
	defer cryptoDomain.Zero(value)

	// Call use case with decoded bytes
	secret, err := h.secretUseCase.CreateOrUpdate(c.Request.Context(), path, value)
	if err != nil {
		httputil.HandleErrorGin(c, err, h.logger)
		return
	}

	// Return response with metadata only (no plaintext)
	response := dto.MapSecretToCreateResponse(secret)
	c.JSON(http.StatusCreated, response)
}

// GetHandler retrieves and decrypts a secret by path, optionally by version.
// GET /v1/secrets/*path?version=N - Requires DecryptCapability.
// Returns 200 OK with plaintext value. SECURITY: Plaintext is zeroed after response.
func (h *SecretHandler) GetHandler(c *gin.Context) {
	// Extract and validate path from URL parameter
	path, ok := h.extractPath(c)
	if !ok {
		return
	}

	var secret *secretsDomain.Secret
	var err error

	// Check for version query parameter
	versionStr := c.Query("version")
	if versionStr != "" {
		version, parseErr := strconv.ParseUint(versionStr, 10, 32)
		if parseErr != nil {
			httputil.HandleBadRequestGin(
				c,
				errors.New("invalid version parameter: must be a positive integer"),
				h.logger,
			)
			return
		}
		if version == 0 {
			httputil.HandleBadRequestGin(
				c,
				errors.New("version must be greater than 0"),
				h.logger,
			)
			return
		}
		secret, err = h.secretUseCase.GetByVersion(c.Request.Context(), path, uint(version))
	} else {
		secret, err = h.secretUseCase.Get(c.Request.Context(), path)
	}

	if err != nil {
		httputil.HandleErrorGin(c, err, h.logger)
		return
	}

	// SECURITY: Zero plaintext after mapping to response
	defer cryptoDomain.Zero(secret.Plaintext)

	// Map to response (includes plaintext value)
	response := dto.MapSecretToGetResponse(secret)
	c.JSON(http.StatusOK, response)
}

// DeleteHandler soft deletes a secret by its path.
// DELETE /v1/secrets/*path - Requires DeleteCapability.
// Returns 204 No Content.
func (h *SecretHandler) DeleteHandler(c *gin.Context) {
	// Extract and validate path from URL parameter
	path, ok := h.extractPath(c)
	if !ok {
		return
	}

	// Call use case
	if err := h.secretUseCase.Delete(c.Request.Context(), path); err != nil {
		httputil.HandleErrorGin(c, err, h.logger)
		return
	}

	// Return 204 No Content with empty body
	c.Data(http.StatusNoContent, "application/json", nil)
}

// ListHandler retrieves secrets with cursor-based pagination support.
// GET /v1/secrets?after_path=/path/to/secret&limit=50 - Requires ReadCapability.
// Returns 200 OK with paginated secret list (excludes plaintext value for security).
// Ordered by path ascending. Uses cursor pagination with after_path parameter.
func (h *SecretHandler) ListHandler(c *gin.Context) {
	// Parse cursor and limit query parameters
	afterPath, limit, err := httputil.ParseStringCursorPagination(c, "after_path")
	if err != nil {
		httputil.HandleBadRequestGin(c, err, h.logger)
		return
	}

	// Call use case with limit + 1 to detect if there are more results
	secrets, err := h.secretUseCase.ListCursor(c.Request.Context(), afterPath, limit+1)
	if err != nil {
		httputil.HandleErrorGin(c, err, h.logger)
		return
	}

	// Determine if there are more results and set next cursor
	var nextCursor *string
	if len(secrets) > limit {
		// More results exist, use the last visible item's path as cursor
		secrets = secrets[:limit]
		cursorValue := secrets[len(secrets)-1].Path
		nextCursor = &cursorValue
	}

	// Map to response
	response := dto.MapSecretsToListResponse(secrets, nextCursor)
	c.JSON(http.StatusOK, response)
}